+++ src/devices/device.ts
@@ -170,7 +170,7 @@ export abstract class deviceBase {
         this.errorLog(`${this.device.deviceClass}: ${this.accessory.displayName} Too Many Requests, statusCode: ${statusCode}, Action: ${action}`)
         break
       case 500:
-        this.errorLog(`${this.device.deviceClass}: ${this.accessory.displayName} Internal Server Error (Meater Server), statusCode: ${statusCode}, Action: ${action}`)
+        this.errorLog(`${this.device.deviceClass}: ${this.accessory.displayName} Internal Server Error (Resideo Server), statusCode: ${statusCode}, Action: ${action}`)
         break
       default:
         this.infoLog(`${this.device.deviceClass}: ${this.accessory.displayName} Unknown statusCode: ${statusCode}, Action: ${action}, Report Bugs Here: https://bit.ly/homebridge-resideo-bug-report`)
@@ -207,7 +207,7 @@ export abstract class deviceBase {
       this.debugLog('The client has exceeded the number of requests allowed for a given time window.')
     } else if (e.message.includes('500')) {
       this.errorLog(`${this.device.deviceClass}: ${this.accessory.displayName} failed to ${action}, Internal Server Error`)
-      this.debugLog('An unexpected error on the SmartThings servers has occurred. These errors should be rare.')
+      this.debugLog('An unexpected error on the Resideo servers has occurred. These errors should be rare.')
     } else {
       this.errorLog(`${this.device.deviceClass}: ${this.accessory.displayName} failed to ${action},`)
     }